Build when deb/* changes
The build calls to Protonmail's repo and builds from it. When that repo changes (i.e. when deb/* changes), the image should be built again.

Currently, that repo changed a few days ago, and the "build" image has not updated in a year. I don't see a reason to never update the image when the upstream repo updates frequently.